+++ CHANGELOG.md
@@ -29,6 +29,8 @@ To release a new version (e.g. from `1.0.0` -> `2.0.0`):
   all cases such that it is consistent when used in the budget optimization that
   optimizes revenue.
 * Fix an error in the data loader that occurs when the geo column is an integer.
+* Add a _check_if_no_time_variation method to Meridian to raise an error if a
+  variable has no time variation.
 
 ## [0.7.0] - 2024-09-20